// src/api/schema.ts
export type LocalizedString = Record<string, string>;

export type MangaResponse = {
    result: 'ok' | 'error'
    /** Default: "entity" */
    response: string
    data: Manga
};

export type ChapterResponse = {
    result: 'ok' | 'error'
    /** Default: "entity" */
    response: string
    data: Chapter
};

export type Relationship = {
    /** UUID formatted string */
    id: string
    type: string
    /** Only present if you are on a Manga entity and a Manga relationship */
    related: 'monochrome' | 'main_story' | 'adapted_from' | 'based_on' | 'prequel' | 'side_story' | 'doujinshi' | 'same_franchise' | 'shared_universe' | 'sequel' | 'spin_off' | 'alternate_story' | 'alternate_version' | 'preserialization' | 'colored' | 'serialization'
    /** If Reference Expansion is applied, contains objects attributes */
    attributes: any | null
};
// ...
export type Tag = {
    /** UUID formatted string */
    id: string
    type: 'tag'
    attributes: TagAttributes
    relationships: Relationship[]
};

export type TagAttributes = {
    name: LocalizedString
    description: LocalizedString
    group: 'content' | 'format' | 'genre' | 'theme'
    /** ```console
     * Minimum: 1
     * ``` */
    version: number
};